#22 Fix performance issue when the attributes contain nulls. Split out tests that deal with the larget dataset to a separate suite so it is easier to execute them separately.
